# docs/swagger.yaml
basePath: /
definitions:
    handler.DriverLocationResponse:
        properties:
            driver_id:
                type: string
            heading:
                type: number
            lat:
                type: number
            lng:
                type: number
            speed:
                type: number
            status:
                type: string
            updated_at:
                type: string
        type: object
    handler.DriverResponse:
        properties:
            driver:
                $ref: '#/definitions/model.Driver'
        type: object
...
    handler.ErrorDetail:
        properties:
            code:
                type: string
            message:
                type: string
        type: object
    handler.ErrorResponse:
        properties:
            error:
                $ref: '#/definitions/handler.ErrorDetail'
        type: object
...
    handler.NearbyDriversResponse:
        properties:
            drivers: {}
        type: object
...
    handler.locationUpdate:
        properties:
            heading:
                type: number
            lat:
                type: number
            lng:
                type: number
            speed:
                type: number
        required:
            - lat
            - lng
        type: object
...
    handler.updateDriverRequest:
        properties:
            first_name:
                type: string
            last_name:
                type: string
            photo_url:
                type: string
        type: object
...
    handler.updateStatusRequest:
        properties:
            status:
                type: string
        required:
            - status
        type: object
...
    model.Driver:
        properties:
            created_at:
                type: string
            first_name:
                type: string
            last_name:
                type: string
            onboarding_status:
                type: string
            photo_url:
                type: string
            rating_summary:
                type: string
            status:
                type: string
            updated_at:
                type: string
            user_id:
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: 'Go + Gin ride-hailing backend: auth, rides, driver dispatch, geolocation, navigation, and places.'
    title: Ride-Hailing API
    version: 0.1.0
paths:
...
    /api/v1/driver/me:
        get:
            operationId: getDriver-me
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.DriverResponse'
                "404":
                    description: Driver not found
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get the driver profile
            tags:
                - driver
        put:
            consumes:
                - application/json
            operationId: putDriver-me
            parameters:
                - description: Profile fields to update
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.updateDriverRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.DriverResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Update the driver profile
            tags:
                - driver
...
    /api/v1/driver/me/status:
        put:
            consumes:
                - application/json
            operationId: putDriver-me-status
            parameters:
                - description: New status (e.g. online, offline)
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.updateStatusRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.DriverResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Update the driver availability status
            tags:
                - driver
...
    /api/v1/driver/register:
        post:
            description: Promotes the current user to the driver role and creates a driver profile.
            operationId: postDriver-register
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/handler.DriverResponse'
            security:
                - BearerAuth: []
            summary: Register the authenticated user as a driver
            tags:
                - driver
...
    /api/v1/drivers/{id}/location:
        get:
            operationId: getDrivers-ByID-location
            parameters:
                - description: Driver ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.DriverLocationResponse'
                "404":
                    description: Driver location not found
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get a driver's current location
            tags:
                - geo
...
    /api/v1/geo/driver/location:
        put:
            consumes:
                - application/json
            description: Upserts the driver position and streams it to the rider over WebSocket.
            operationId: putGeo-driver-location
            parameters:
                - description: Location update
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.locationUpdate'
            produces:
                - application/json
            responses:
                "204":
                    description: No content
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Update failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Update a driver's live location
            tags:
                - geo
    /api/v1/geo/driver/location/batch:
        put:
            consumes:
                - application/json
            operationId: putGeo-driver-location-batch
            parameters:
                - description: List of location updates
                  in: body
                  name: body
                  required: true
                  schema:
                    items:
                        $ref: '#/definitions/handler.locationUpdate'
                    type: array
            produces:
                - application/json
            responses:
                "204":
                    description: No content
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Batch update a driver's live location
            tags:
                - geo
...
    /api/v1/geo/nearby-drivers:
        get:
            description: Returns online drivers within a radius of the given coordinates.
            operationId: getGeo-nearby-drivers
            parameters:
                - description: Latitude
                  in: query
                  name: lat
                  required: true
                  type: number
                - description: Longitude
                  in: query
                  name: lng
                  required: true
                  type: number
                - default: 5000
                  description: Search radius in meters (default 5000)
                  in: query
                  name: radius
                  type: number
                - default: 20
                  description: Maximum results (max 50)
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.NearbyDriversResponse'
                "422":
                    description: Invalid parameters
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Query failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Find nearby drivers
            tags:
                - geo
    /api/v1/geo/rider/location:
        put:
            consumes:
                - application/json
            operationId: putGeo-rider-location
            parameters:
                - description: Location update
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.locationUpdate'
            produces:
                - application/json
            responses:
                "204":
                    description: No content
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Update a rider's live location
            tags:
                - geo
...
securityDefinitions:
    BearerAuth:
        description: 'JWT access token. Include as `Authorization: Bearer <token>`.'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"